+++ awb_generic_sms/__manifest__.py
@@ -21,7 +21,7 @@
 
     'category': 'Sms Integration',
 
-    'version': '13.0.1.3.7',
+    'version': '13.0.1.3.8',
 
     'depends': [
         'base',