+++ awb_payment_allocation/__manifest__.py
@@ -14,16 +14,11 @@
     'author': "Achieve Without Borders, Inc",
     'website': "http://www.achievewithoutborders.com",
     'category': 'Accounting/Accounting',
-    'version': '13.0.1.0.3',
+    'version': '13.0.1.0.1',
     'depends': ['account'],
     'data': [
         'security/ir.model.access.csv',
-        'views/account_payment_template.xml',
         'views/account_payment_view.xml',
-        'views/payment_allocation_view.xml',
-    ],
-    'qweb': [
-        "static/src/xml/account_reconciliation_view.xml",
     ],
     'installable': True,
     'application': False,