+++ awb_payment_allocation/models/__init__.py
@@ -7,4 +7,3 @@
 
 # from . import account
 from . import account_payment
-from . import payment_allocation