+++ awb_subscriber_bill/__manifest__.py
@@ -21,7 +21,7 @@
 
     'category': 'Localization',
 
-    'version': '13.0.1.7.3',
+    'version': '13.0.1.7.4',
 
     'depends': ['account_accountant', 'sale_subscription'],