+++ awb_subscriber_bill_automation/models/subscription.py
@@ -46,7 +46,7 @@ class SaleSubscription(models.Model):
 
     # New Fields
 
-    plan_type = fields.Many2one('product.plan.type', compute='_compute_plan_type')
+    plan_type = fields.Many2one('product.plan.type', compute='_compute_plan_type', store=True)
 
     # Business Logic
 
@@ -112,7 +112,10 @@ class SaleSubscription(models.Model):
             self.env['sale.subscription'].provision_and_activation(self.record, main_plan, last_subscription, last_subs_main_plan, ctp)
 
             # Helper to update Odoo Opportunity
-            # self._update_account(main_plan, self.record, sf_update_type, max_fail_retries)            
+            try:
+                self._update_account(main_plan, self.record, sf_update_type, max_fail_retries)
+            except:
+                _logger.error(f'SMS:: !!! Failed SF Account Update - Subscription code {self.record.code}')            
 
         if not ctp:
             self.env['sale.subscription'].generate_atmref(self.record, max_fail_retries)
@@ -144,7 +147,7 @@ class SaleSubscription(models.Model):
         _logger.debug(f'SMS:: Customer Number: {customer_id}')
 
         last_subscription = False
-        subscriptions = self.env['sale.subscription'].search([('customer_number','=', customer_id)], order='id desc', limit=2)
+        subscriptions = self.env['sale.subscription'].search([('customer_number','=', customer_id),('plan_type','=', plan_type)], order='id desc', limit=2)
         
         if len(subscriptions) == 2:
             last_subscription = subscriptions[1]
@@ -504,9 +507,6 @@ class SaleSubscription(models.Model):
     def _get_datetime_now(self):
         _logger.info(f'SMS:: function: _get_datetime_now')    
         try:
-            # timezone = self.env.user.tz or pytz.utc
-            # now = datetime.datetime.now(pytz.timezone(timezone))
-
             # Current time in UTC
             now_utc = datetime.now(timezone('UTC'))
             # Convert to Asia/Manila time zone