+++ streamtech_salesforce_connector/__manifest__.py
@@ -15,7 +15,7 @@
     'author': "Achieve Without Borders, Inc",
     'website': "http://www.achievewithoutborders.com",
     'category': 'Sales',
-    'version': '13.0.1.2.4',
+    'version': '13.0.1.2.6',
     'depends': ['salesforce_connector', 'awb_subscriber_bill_automation', 'awb_subscriber_product_information'],
     'data': [
         'security/ir.model.access.csv',