+++ awb_collection_reports/CHANGELOG.md
@@ -19,3 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [13.0.1.0.2] - 2021-09-07
 
 - [FIX] Daily collection report fixes
+
+## [13.0.1.0.3] - 2021-09-24
+
+- [FIX] Daily collection report hotfix