+++ awb_subscriber_bill/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [v13.0.1.7.4] - 2021-08-27
+    - [FIX] Removed the changes for Send Billing SMS fix
+
 ## [v13.0.1.7.3] - 2021-08-23
 
     - [FIX] Minimum Due Currency